AI Studio functionality improvements (#62)
* shrink Generative AI empty placholder to a minimum

* improve version bumping of the css/js files

* change included css/js versioning system

* fix sizing of the AI placeholder, and fix display of circle images in confimration dialogs

* updated csum for twbuild.css

* implement multiselect actions

* bump css/js versions

* make cancel not cancel the selection

* swap move to folder icon

* close media action menu only when no dialog is open

* prevent dialog from closing when action is in progress

* add IOU placeholders

* swap profile edit and logout icons

* swap Generative AI icon

* fix z values for bars and dialogs

+++ build/bump_css_js_version.py
@@ -1,76 +1,65 @@
 #!/usr/bin/env python3
+
 import os
 import re
 import argparse
+import hashlib
 
+def calculate_checksum(file_path):
+    with open(file_path, 'rb') as f:
+        content = f.read()
+        checksum = hashlib.md5(content).hexdigest()
+    return checksum
 
-def update_css(m):
-    prefix = m.group(1)
-    href = m.group(2)
-    version = m.group(4)
-    quote = m.group(5)
+def update_version(m, root_folder):
+    tag = m.group(1)
+    path = m.group(2)
 
-    if "http:" in href or "https:" in href:
+    if "http:" in path or "https:" in path:
         return m.group(0)  # return the entire match as is
 
-    if version:
-        new_version = int(version) + 1
+    file_path = os.path.join(root_folder, path.lstrip('/'))
+    if os.path.isfile(file_path):
+        checksum = calculate_checksum(file_path)
+        return f'{tag}{path}?v={checksum}"'
     else:
-        new_version = 1
-    return f'{prefix}{href}?v={new_version}{quote}'
+        return m.group(0)  # return the entire match as is if file not found
 
-
-def update_js(m):
-    prefix = m.group(1)
-    src = m.group(2)
-    version = m.group(4)
-    quote = m.group(5)
-
-    if "http:" in src or "https:" in src:
-        return m.group(0)  # return the entire match as is
-
-    if version:
-        new_version = int(version) + 1
-    else:
-        new_version = 1
-    return f'{prefix}{src}?v={new_version}{quote}'
-
-
-def update_version(file_path):
+def process_file(file_path, root_folder):
     with open(file_path, 'r+') as f:
         content = f.read()
         f.seek(0)
 
-        # Update CSS references
-        updated_content = re.sub(r'(<link [^>]*?href=["\'])([^"\']*\.css)(\?v=([0-9]+))?(["\'])',
-                                 update_css, content)
+        # Update CSS and JS references
+        updated_content = re.sub(r'(<(?:link|script)[^>]*(?:href|src)=")([^"]*?\.(?:css|js))(?:\?v=[^"]*)?(")',
+                                 lambda m: update_version(m, root_folder), content)
 
-        # Update JS references
-        updated_content = re.sub(r'(<script [^>]*?src=["\'])([^"\']*\.js)(\?v=([0-9]+))?(["\'])',
-                                 update_js, updated_content)
+        if updated_content != content:
+            print(f'Updated CSS and JS checksums for {file_path}')
+            f.write(updated_content)
+            f.truncate()
 
-        f.write(updated_content)
-        f.truncate()
+def find_and_update_files(start_path='.', skip_folders=[], root_folder=None):
+    if root_folder is None:
+        root_folder = start_path
 
-
-def find_and_update_files(start_path='.', skip_folders=[]):
-    for root, _, filenames in os.walk(start_path):
-        # Skip folders that are in skip_folders list
-        if any(skip in root for skip in skip_folders):
-            continue
+    for root, dirs, filenames in os.walk(start_path):
+        dirs[:] = [d for d in dirs if d not in skip_folders]
 
         for filename in filenames:
             if filename.endswith('.php'):
                 full_path = os.path.join(root, filename)
-                update_version(full_path)
-
+                process_file(full_path, root_folder)
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(
-        description='Update CSS and JS version numbers in PHP files.')
-    parser.add_argument('--start', default='.',
+        description='Update CSS and JS version numbers in PHP files using file checksums.')
+    parser.add_argument('--start', default='..',
                         help='Starting directory for the search.')
-    parser.add_argument(
-        '--skip', nargs='*', default=['vendor', 'libs', 'scripts'], help='List of folders to skip.')
+    parser.add_argument('--skip', nargs='*', default=['vendor', 'libs', 'scripts', 'build'],
+                        help='List of folders to skip.')
+    parser.add_argument('--root', default='..',
+                        help='Root folder from which to find target files.')
     args = parser.parse_args()
-    find_and_update_files(args.start, args.skip)
+
+    find_and_update_files(args.start, args.skip, args.root)
